requirements.txt
ipykernel
selenium
pandas
pyyaml
flask